+++ OgreMain/src/OgreETCCodec.cpp
@@ -355,7 +355,7 @@ namespace Ogre {
         if (header.glType == 0 || header.glFormat == 0)
             imgData->flags |= IF_COMPRESSED;
 
-        size_t numFaces = header.numberOfFaces;
+        uint32 numFaces = header.numberOfFaces;
         if (numFaces > 1)
             imgData->flags |= IF_CUBEMAP;
         // Calculate total size from number of mipmaps, faces and size
